Use double DurationMs; add CPU time deltas
Change pipeline duration to double (sub-millisecond precision) and propagate the type across IPipelineMetrics, PipelineMetrics, collector, logging sink and tests. Add high-resolution node start timestamp (Stopwatch) and use Stopwatch.GetElapsedTime to compute node durations. Collect best-effort per-node CPU time deltas using Process.TotalProcessorTime (store initial value on node start, compute delta on node end) and expose it to the collector. Update LoggingPipelineMetricsSink to log DurationMs with fractional precision and adjust formatting. Update AutoObservabilityScope to use TotalMilliseconds for averages. Update docs and README to reflect CPU metric semantics and fractional duration support.

+++ src/NPipeline/Observability/Metrics/IPipelineMetrics.cs
@@ -66,7 +66,7 @@ public interface IPipelineMetrics
     /// <summary>
     ///     The total duration of the pipeline execution in milliseconds.
     /// </summary>
-    long? DurationMs { get; }
+    double? DurationMs { get; }
 
     /// <summary>
     ///     Whether the pipeline execution was successful.
